pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctxos"
version = "1.0.1"
description = "Debian Base Kit - Distribution Factory"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}

dependencies = [
    "pywebview==6.1",
    "pydbus==0.6.0",
    "PyGObject==3.46.0",
    "python-apt==2.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pylint==3.0.3",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "black==23.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["backend*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pylint."messages control"]
disable = ["C0111", "C0103"]  # Missing docstring, invalid names
